+++ mex/grippeweb/main.py
@@ -12,6 +12,7 @@ from mex.common.models import (
     ExtractedPerson,
     ExtractedPrimarySource,
     ExtractedResource,
+    ExtractedVariable,
     ExtractedVariableGroup,
 )
 from mex.common.primary_source.transform import (
@@ -33,6 +34,8 @@ from mex.grippeweb.settings import GrippewebSettings
 from mex.grippeweb.transform import (
     transform_grippeweb_access_platform_to_extracted_access_platform,
     transform_grippeweb_resource_mappings_to_extracted_resources,
+    transform_grippeweb_variable_group_to_extracted_variable_groups,
+    transform_grippeweb_variable_to_extracted_variables,
 )
 from mex.mapping.extract import extract_mapping_data
 from mex.pipeline import asset, run_job_in_process
@@ -79,6 +82,15 @@ def grippeweb_resource_mappings() -> list[dict[str, Any]]:
         extract_mapping_data(file, ExtractedResource)
         for file in Path(settings.mapping_path).glob("resource_*.yaml")
     ]
+
+
+@asset(group_name="grippeweb")
+def grippeweb_variable() -> dict[str, Any]:
+    """Extract Grippeweb `variable` default values."""
+    settings = GrippewebSettings.get()
+    return extract_mapping_data(
+        settings.mapping_path / "variable.yaml", ExtractedVariable
+    )
 
 
 @asset(group_name="grippeweb")
@@ -187,6 +199,45 @@ def grippeweb_extracted_resource_dict(
     )
     load(list(extracted_resources.values()))
     return extracted_resources
+
+
+@asset(group_name="grippeweb")
+def grippeweb_extracted_variable_group(
+    grippeweb_variable_group: dict[str, Any],
+    grippeweb_columns: dict[str, dict[str, list[Any]]],
+    grippeweb_extracted_resource_dict: dict[str, ExtractedResource],
+    extracted_primary_source_grippeweb: ExtractedPrimarySource,
+) -> list[ExtractedVariableGroup]:
+    """Transform Grippeweb values to extracted variable groups and load to sinks."""
+    extracted_variable_groups = (
+        transform_grippeweb_variable_group_to_extracted_variable_groups(
+            grippeweb_variable_group,
+            grippeweb_columns,
+            grippeweb_extracted_resource_dict,
+            extracted_primary_source_grippeweb,
+        )
+    )
+    load(extracted_variable_groups)
+    return extracted_variable_groups
+
+
+@asset(group_name="grippeweb")
+def grippeweb_extracted_variable(
+    grippeweb_variable: dict[str, Any],
+    grippeweb_extracted_variable_group: list[ExtractedVariableGroup],
+    grippeweb_columns: dict[str, dict[str, list[Any]]],
+    grippeweb_extracted_resource_dict: dict[str, ExtractedResource],
+    extracted_primary_source_grippeweb: ExtractedPrimarySource,
+) -> None:
+    """Transform Grippeweb default values to extracted variables and load to sinks."""
+    extracted_variables = transform_grippeweb_variable_to_extracted_variables(
+        grippeweb_variable,
+        grippeweb_extracted_variable_group,
+        grippeweb_columns,
+        grippeweb_extracted_resource_dict,
+        extracted_primary_source_grippeweb,
+    )
+    load(extracted_variables)
 
 
 @entrypoint(GrippewebSettings)